+++ src/inputscope/conf.py
@@ -36,7 +36,7 @@ import sys
 
 """Program title, version number and version date."""
 Title = "InputScope"
-Version = "1.11.dev0"
+Version = "1.11.dev1"
 VersionDate = "10.04.2024"
 
 """TCP port of the web user interface."""