+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 CHANGELOG
 =========
 
-1.11, 2024-04-10
+1.11, 2024-04-11
 ----------------
+- scale mouse heatmap to screen height
 - make keyboard heatmap image path configurable
 - fix logging Numpad operator keys in Linux (issue #30)
 - fix SQLite-related warnings in Python 3.12+ (issue #30)